ball

+++ bot/cogs/extra/media.py
@@ -52,7 +52,7 @@ class mediastuff(commands.Cog):
 
         dlpath = "'temp/%(id)s.%(ext)s'"
         usage = f"{env.prefix}ytdl 'https://youtube.com/...' video / audio (default: video)"
-        if not query:
+        if not query or query == None:
             await ctx.reply(f"No YouTube URL found.\n```{usage}```")
         else:
             # User download format options, 'video' is the default.
@@ -74,43 +74,51 @@ class mediastuff(commands.Cog):
                 opts = f'-x --audio-format {media_format}'
             else:
                 await ctx.reply(f"Invalid usage.\n```{usage}```")
+                opts = None
 
-            msg = await ctx.reply(f"Downloading <{query}>...")
-            if not os.path.exists('temp'):
-                os.mkdir('temp')
+            #if query.startswith("https://youtu") == False:
+            #    await ctx.reply("Not a YouTube URL. Only YouTube videos are allowed to be downloaded.")
+            #else:
+            if query.startswith('<') and query.endswith('>'):
+                query = query.replace('<', '').replace('>', '')
+            if opts != None:
+                msg = await ctx.reply(f"Downloading <{query}>...")
+                if not os.path.exists('temp'):
+                    await aiofiles.os.mkdir('temp')
 
-            # Extract video info for the ID.
-            with yt_dlp.YoutubeDL() as ydl:
-                info_dict = ydl.extract_info(query, download=False)
-                video_id = info_dict.get("id", None)
-            # Run yt-dlp to download the video & convert if needed
-            # a pretty unsafe way of doing it, but yt_dlp doesn't support asyncio (and i also dont know async multiprocessing)
-            try:
-                subprocesspipe = asyncio.subprocess.PIPE
-                child = await asyncio.create_subprocess_shell(
-                    f"python3 -m yt_dlp --max-filesize 350M -f 22/18/5/36 --no-playlist {opts} --break-on-existing --no-exec -o {dlpath} {query}",
-                    stderr=subprocesspipe,
-                    stdout=subprocesspipe
-                )
-                stdout, stderr = await child.communicate()
-            except Exception as e:
-                await msg.edit(f"```{e}```")
-            # Grab the exit code to know if yt-dlp succeeded or not
-            exitcode = child.returncode
-            vidfile = f"{video_id}.{media_format}"
-            tmplocation = f"temp/{vidfile}"
-            await msg.delete()          # Delete original download msg
-            # If yt-dlp ran successfully and the file exists, upload it.
-            try:
-                if exitcode == 0 and os.path.exists(tmplocation):
-                    await ctx.reply(file=discord.File(tmplocation))
-                    await asyncio.sleep(14400)
-                    await aiofiles.os.remove(tmplocation)
-            except:
-                await msg.reply(f"Something went wrong. The video download likely failed and no file was found to upload. ```{exitcode}```")
-            else:
-                await msg.reply(f"Something has gone abnormally wrong, and I don't know the error.")
-            # remove the file after a certain amount of time (in this case, 14400 seconds is 4 hours)
+                # Extract video info for the ID.
+                with yt_dlp.YoutubeDL() as ydl:
+                    info_dict = ydl.extract_info(query, download=False)
+                    video_id = info_dict.get("id", None)
+                # Run yt-dlp to download the video & convert if needed
+                # a pretty unsafe way of doing it, but yt_dlp doesn't support asyncio (and i also dont know async multiprocessing)
+                async with ctx.typing():
+                    try:
+                        subprocesspipe = asyncio.subprocess.PIPE
+                        child = await asyncio.create_subprocess_shell(
+                            f"python3 -m yt_dlp --max-filesize 350M -f 22/18/5/36 --no-playlist {opts} --break-on-existing --no-exec -o {dlpath} {query}",
+                            stderr=subprocesspipe,
+                            stdout=subprocesspipe
+                        )
+                        stdout, stderr = await child.communicate()
+                    except Exception as e:
+                        await msg.edit(f"```{e}```")
+                    # Grab the exit code to know if yt-dlp succeeded or not
+                exitcode = child.returncode
+                vidfile = f"{video_id}.{media_format}"
+                tmplocation = f"temp/{vidfile}"
+                await msg.delete()          # Delete original download msg
+                # If yt-dlp ran successfully and the file exists, upload it.
+                try:
+                    if exitcode == 0 and os.path.exists(tmplocation):
+                        await ctx.reply(file=discord.File(tmplocation))
+                        await asyncio.sleep(14400)
+                        await aiofiles.os.remove(tmplocation)
+                    else:
+                        await msg.reply(f"Something went wrong. The video download likely failed (and/or was too big) ```{exitcode}```")
+                except:
+                    await msg.reply(f"Something has gone wrong, and I don't know the error.")
+                # remove the file after a certain amount of time (in this case, 14400 seconds is 4 hours)
 
 
 def setup(bot: commands.Bot):
